MPC_Launch: Remove not supported PMIX_LOCAL_PEERS
The pmix used by the CI doesn't seem to support the PMIX_LOCAL_PEERS
constants as a get to it returns a PMIX_ERR_NOT_FOUND (-46).
This commit fix this issue by retrieving the node layout computed in the
intialization in order to translate the local rank to a global one.

Note this translation is only needed when N > 1 and N < n, therefore the
initial `mpc_launch_pmi_get` function can be used as a shortcut to avoid
this translation. It would mean a diminution of the duplicate code as
well.