MPC_Launch: fix wrong termination with PMIx on abortion
Similarly to !238, the termination of the calling processes may occur
before the PMIx server is notified. To mitigate the issue and keep the
intended behaviour, the caller waits now for the PMIx server sigterm.
If after 5s, the signal doesn't arrive, `pmix_finalize` is called before
calling the abort to ensure that the PMIx server is notified.

Closes #210