chore(mpc_c*): Add -h as a shorthand option for --help
The `-h` option was not recognized as an option which leads to odd
behavior. For instance, launching `mpc_cc -h` results in a cc1 error
pointing that it was unable to find mpc.h.

We could discuss the implementation of a short help which will be shown
by this new option to differentiate it from the long form `--help`. It
is not the case for now the two options are strictly identical.