fix(mpc_compiler_common): Activate privatization only if possible
The default behaviour is to want the privatization. Nevertheless, if it
was not possible (the compiler isn't enable to), we still tried to do
so. This commit splits the concerns into two states: do we want
privatization and can we privatize. Both have to be true to actually
activate privatization.

Closes #212