fix(mpc_print_config): Add absolute path support to exename
The default behaviour is to set different defaults for the parameters on
whether the program have to run something or if it is only requested to
print something. However, the check is made on the name of the
executable and this one was not translated from possible absolute path
to relative path. This is an issue as the invocation of `mpc_print_config`
in `mpcrun` is done with the absolute path which leads to the wrong
behaviour to be adopted. This commit adds a translation step to ensure
that calling `mpc_print_config` with its absolute path or a relative one
will produce the same result.