fix(CI): remove cyclic links creation and unnecessary spack loads
The linkage to the MPC dependencies directory was run for each build.
With the omission of the `-n` option, ln was creating a circular
symbolic link. To avoid this issue (and stop modifying files outside the
runner node), this step is run only once in the Env Sanitization step.

The loading of spack modules is unnecessary for the builds job!
Furthermore, it provokes failure on builds with the newly created spack
environment containing llvm compilers.