fix(CI): rework extended tests to always run all the tests
The benchmarks are now run a singular PCVS command to avoid that a
failing benchmark stop the other from running.

A variable (BENCHMARKS_BRANCH) has been added to target the branch of
PCVS--Benchmarks to use. The default remains the same, targeting the
devel branch which should be the most up-to-date.
Its value should be printed out on the log to ease debugging

New stages has been added to split the extended tests into categories
(threading or network).

PCVS runs with all the output (rather than only errors) is activated for some benchmarks (OSU and
RMA-MT). It is nice to see the results of performance testing in numbers
and not just as functional tests.