fix(MPI): change error-handling policy for mpit and fix error printing
The error printing in case of unknown MPI error was erroneous. It
lacked a parameter resulting in an invalid read in memory.
It has been fixed and the MPI_T_ERR_INVALID_NAME has been added to the
list.

Some query functions in MPI Tools has been dial down in terms of
error-handling. Before this change, those functions would abort on error.
Now, they return the appropriated error-code to the user as querying a
missing object should not be a fatal error.
It was prone to crash in circumstances where the query functions were
used to find if an object already exist before creating it if not.
It is a scheme found in the OSU micro benchmarks.