feat: add specific error-code abortion support
The `mpc_common_debug_abort` helper function now takes in an error-code
which is used underneath to exit with a proper error-code.
The default behaviour was to exit with 6 as an error-code and it is still
the preferred way in function with no specification of a precise
error-code.
It allows the `MPI_Abort` to be more compliant with the standard by
returning the passed error-code instead of ignoring it.