feat(LCP): do not initialize shm if only one process per node (#231)
We filter the components and do not load shm ones if only one process
per node is requested. It should result in a lower memory footprint and
a faster startup time for pure internode jobs.

The number of nodes and process is not retrieved via the user passed
arguments but the internal representation of the topology gotten by
PMI(x).

Closes #231